+++ src/main/java/htl/steyr/repus_oiram_klichtl_rhaubner_mbrambe_lmikota/GameElements/FloorEnemy.java
@@ -9,7 +9,7 @@ public class FloorEnemy extends Enemy implements Runnable {
     private boolean isBlockUnderIt;
     private double enemyGravity = 4;
 
-    public FloorEnemy(Image enemyImage, int enemySize, int tileSize, int enemyMovementX, Player player, int map[][], double xCord, double yCord) {
+    public FloorEnemy(Image enemyImage, int enemySize, int tileSize, double enemyMovementX, Player player, int map[][], double xCord, double yCord) {
         super(enemyImage, enemySize, tileSize, enemyMovementX, player, xCord, yCord);
         this.map = map;
     }
