+++ src/main/java/htl/steyr/repus_oiram_klichtl_rhaubner_mbrambe_lmikota/GameElements/SkyEnemy.java
@@ -4,7 +4,7 @@ import javafx.application.Platform;
 import javafx.scene.image.Image;
 
 public class SkyEnemy extends Enemy implements Runnable {
-    public SkyEnemy(Image enemyImage, int enemySize, int tileSize, int enemyMovementX, Player player, double xCord, double yCord) {
+    public SkyEnemy(Image enemyImage, int enemySize, int tileSize, double enemyMovementX, Player player, double xCord, double yCord) {
         super(enemyImage, enemySize, tileSize, enemyMovementX, player, xCord, yCord);
     }
